pyproject.toml
[project]
name = "miga"
version = "1.0.0"
description = "MCP Intelligence Gateway Architecture — Unified Cisco AI/ML Platform"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [{name = "Keenan Williams"}]
keywords = ["mcp", "cisco", "ai", "ml", "network", "automation", "agntcy"]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "redis[hiredis]>=5.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "click>=8.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
infer = ["pandas>=2.1.0", "scipy>=1.11.0", "scikit-learn>=1.3.0"]
dev = ["ruff>=0.5.0", "pytest>=8.0.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0.0"]

[project.scripts]
miga-cli = "packages.cli.miga_cli:cli"

[tool.setuptools.packages.find]
include = ["miga_shared*", "packages*", "servers*"]

[tool.ruff]
target-version = "py311"
line-length = 100
[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
